refactor: move kubectl-cp to staging
