+++ staging/src/k8s.io/kubectl/pkg/cmd/cp/cp_test.go
